refactor custom buttons to use singular custom button component
